06_kafka/requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart
pymongo
redis
confluent-kafka


# pandas
# prometheus_client
# requests
# pika
